Create input component with timed scroll effect
