+++ src/apps/chatter/components/chatbox/chatbox.js
@@ -1,11 +1,11 @@
-import { useState, useEffect, useRef } from 'react'
+import { useEffect, useRef } from 'react'
 
 import { Message } from '/elements/message'
 import { useRoom } from '/apps/chatter/hooks/use-room'
+import { MessageInput } from '../message-input'
 import './chatbox.less'
 
 export const ChatBox = ({ roomId }) => {
-  const [textInput, setTextInput] = useState('')
   const [messages, sendNewMessage] = useRoom(roomId)
   const inputEl = useRef(null)
 
@@ -13,46 +13,17 @@ export const ChatBox = ({ roomId }) => {
     inputEl.current && inputEl.current.focus()
   }, [messages])
 
-  const handleSend = () => {
+  const handleSend = (text) => {
     console.log('sending')
-    sendNewMessage(textInput)
-    setTextInput('')
-  }
-
-  const handleKeyUp = (e) => {
-    if (e.keyCode === 13) {
-      handleSend()
-    }
-  }
-
-  const handleInputChange = (e) => {
-    setTextInput(e.target.value)
+    sendNewMessage(text)
   }
 
   return (
     <div className='chatbox'>
-
       {messages.map(Message)}
-
-      <div className='chatbox__input-row'>
-        <input
-          id='chatbox__message-input'
-          ref={inputEl}
-          type='text'
-          size='100'
-          value={textInput}
-          onChange={handleInputChange}
-          onKeyUp={handleKeyUp}
-        />
-
-        <button
-          className='chatbox__send-button btn'
-          onClick={handleSend}
-          disabled={textInput === ''}
-        >
-          Send
-        </button>
-      </div>
+      <MessageInput
+        onSend={handleSend}
+      />
     </div>
   )
 }
