+++ src/apps/chatter/components/message-input/index.js
@@ -1,0 +1,1 @@
+export { MessageInput } from './message-input'
